pyproject.toml
[tool.poetry]
name = "pyriksdagen"
version = "0.8.2rc2"
description = "Access the Riksdagen corpus"
authors = ["ninpnin <vainoyrjanainen@icloud.com>"]
repository = "https://github.com/welfare-state-analytics/riksdagen-corpus"
readme = "README.md"
license = "MIT"
packages = [
    { include = "pyriksdagen", from = "" }
]

[tool.poetry.dependencies]
python = "^3.7"
pyparlaclarin = "*"
pandas = "*"
base58 = "*"
kblab-client = "*"
progressbar2 = "*"
lxml = "*"
unidecode = "*"
dateparser = "*"
beautifulsoup4 = "*"
requests = "*"
numpy = "*"
SPARQLWrapper = "*"
tqdm = "*"
matplotlib = "*"
importlib_resources = "*"
nltk = "*"

[tool.poetry.dev-dependencies]
devtools = "^0.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
